filtro di kalman sbaglia nella previsione, gli array da pushare nei vector devono essere allocati in heap altrimenti l'area di memoria viene sovrascritta

+++ t_tracks.h
@@ -8,7 +8,7 @@ class t_tracks
 {
 public:
 	int id;
-	int* bbox;
+	double* bbox;
 	KalmanFilter kalmanFilter;
 	int age;
 	int totalVisibleCount;
